Add OCI registry client as second fallback
OCI registries don't provide a way to retrieve metadata in the same call
as listing tags.

This means that you have to do a separate API request for each tag if
you want the metadata, which could be very slow for a large number of
tags.

As such, preferably use the selfhosted client as a fallback, but add a
fallback fallback to the oci handler for registries that are
incompatible with the selfhosted API implementation.