OCI: Fetch image metadata for highest version
OCI repos don't provide a way to fetch the metadata alongside the image
tags. Each metadata is a separate API call per tag, and so it could be
very slow to retrieve all the metadata if a image has a lot of tags.

We only really care about the highest version, so just retrieve metadata
for that one.