Revert "enable proxy for remaining clients (#152)"
This reverts commit 293835ed7d81111f4bb7b9096b5e954a14d79c03.

It's unnecessary to specify ProxyFromEnvironment in these cases.
It was required for the selfhosted client, as we are overriding the
default Transport object.

In these cases, Transport are nil, and so should default to
DefaultTransport: https://cs.opensource.google/go/go/+/refs/tags/go1.22.0:src/net/http/client.go;l=60

This already has ProxyFromEnvironment as the default behaviour:
https://github.com/golang/go/blob/master/src/net/http/transport.go#L44

As this is unnecessary, it's better to leave it undefined to ensure that
the other characteristics of DefaultTransport are retained.